Fuse fmpz_mpoly/compose*
